fix colors. add python-specific stuff to gitignore

+++ main.py
@@ -20,7 +20,7 @@ from scipy.cluster.hierarchy import fclusterdata
 
 
 def draw(data):
-    data = data.reshape(n_rows, n_columns)
+    data = data.reshape(n_rows, n_columns).astype(int)
 
     clusters_amount = len(set(data.flatten())) + 1
